backend/foodgram/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
Pillow==8.3.1
PyJWT==2.1.0
requests==2.26.0
djoser==2.1.0 
djangorestframework-simplejwt==4.8.0
django-filter==2.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
asgiref==3.2.10
pytz==2020.1
sqlparse==0.3.1 
python-dotenv==0.19.0
